feat: add Node API for packaging
